pyproject.toml
[tool.poetry]
name = "price-bot"
version = "0.1.0"
description = ""
authors = ["Ilya Dashkevich <94009854+goftok@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.13,<4.0"
requests = "^2.31.0"
rich = "^13.6.0"
python-dotenv = "^1.0.0"
deep-translator = "^1.11.4"
geopy = "^2.4.1"
art = "^6.1"
pre-commit = "^3.7.0"
pytest = "^8.2.1"
brotli = "^1.1.0"
sqlalchemy = "^2.0.36"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
